Moved BasicPanel to top level
