+++ src/main/java/me/jsedwards/configserver/ServerConfigPanel.java
@@ -33,12 +33,18 @@ public class ServerConfigPanel extends JPanel implements Card {
         serverNameLbl.setFont(Main.MAIN_FONT);
         this.add(serverNameLbl, new GridBagConstraints(2, 1, 1, 1, 1, 0, GridBagConstraints.NORTHWEST, GridBagConstraints.HORIZONTAL, new Insets(30, 0, 0, 0), 0, 0));
         // Tabbed pane
-        BasicPanel basicPanel = new BasicPanel();
+        BasicPanel basicPanel = new BasicPanel(this);
         tabbedPane.add("General", basicPanel);
         for (AdvancedPanel panel : advancedPanels) {
             tabbedPane.add(panel.name, panel);
         }
         this.add(tabbedPane, new GridBagConstraints(1, 2, 2, 1, 1, 1, GridBagConstraints.NORTH, GridBagConstraints.BOTH, new Insets(10, 10, 10, 10), 0, 0));
+    }
+
+    public void optimiseConfigs(int sliderValue) {
+        for (AdvancedPanel panel : advancedPanels) {
+            panel.optimise(sliderValue);
+        }
     }
 
     public void setServer(String serverName) {
@@ -64,37 +70,6 @@ public class ServerConfigPanel extends JPanel implements Card {
     public void exit() {
         for (AdvancedPanel panel : advancedPanels) {
             panel.saveProperties();
-        }
-    }
-
-    private class BasicPanel extends JPanel {
-
-        public BasicPanel() {
-            super();
-            this.setLayout(new GridBagLayout());
-            // General optimisation slider
-            JSlider slider = createSlider();
-            this.add(slider, new GridBagConstraints(1, 1, 2, 1, 1, 0, GridBagConstraints.NORTH, GridBagConstraints.HORIZONTAL, new Insets(0, 0, 0, 0), 0, 0));
-            // Slider labels
-            JLabel lowerLbl = new JLabel("More performance");
-            lowerLbl.setHorizontalAlignment(SwingConstants.LEFT);
-            this.add(lowerLbl, new GridBagConstraints(1, 2, 1, 1, 1, 0, GridBagConstraints.WEST, GridBagConstraints.HORIZONTAL, new Insets(0, 0, 0, 0), 0, 0));
-            JLabel upperLbl = new JLabel("Better experience");
-            upperLbl.setHorizontalAlignment(SwingConstants.LEFT);
-            this.add(upperLbl, new GridBagConstraints(2, 2, 1, 1, 0, 0, GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, new Insets(0, 0, 0, 0), 0, 0));
-            // Padding
-            this.add(new JPanel(), new GridBagConstraints(1, 3, 2, 1, 1, 1, GridBagConstraints.NORTH, GridBagConstraints.BOTH, new Insets(0, 0, 0, 0), 0, 0));
-        }
-
-        private JSlider createSlider() {
-            JSlider slider = new JSlider(0, 100);
-            slider.addChangeListener(e -> {
-                int sliderValue = slider.getValue();
-                for (AdvancedPanel panel : ServerConfigPanel.this.advancedPanels) {
-                    panel.optimise(sliderValue);
-                }
-            });
-            return slider;
         }
     }
